src/main.py
#!/usr/bin/env python3

import os
import yaml
import logging
from datetime import datetime, timedelta
import pytz
import requests
from typing import Dict, List, Optional

# Configure logging
logging.basicConfig(
    level=logging.INFO,
    format='%(asctime)s - %(name)s - %(levelname)s - %(message)s'
)
logger = logging.getLogger(__name__)

class CalendarHandler:
    def __init__(self):
        self.timezone = "Europe/Amsterdam"

    def get_todays_events(self) -> List[Dict]:
        """Get today's calendar events."""
        now = datetime.now(pytz.timezone(self.timezone))
        start = now.replace(hour=0, minute=0, second=0, microsecond=0)
        end = start + timedelta(days=1)
        
        # We'll implement the actual calendar query here
        return []

    def analyze_calendar(self) -> Dict:
        """Analyze calendar events and return a summary."""
        events = self.get_todays_events()
        
        # We'll implement the analysis logic here
        return {
            "meeting_density": "light",
            "special_events": [],
            "location_changes": []
        }

class WeatherHandler:
    def __init__(self, api_key: str):
        self.api_key = api_key
        self.base_url = "http://api.openweathermap.org/data/2.5/weather"

    def get_weather(self, location: str) -> Dict:
        """Get current weather for location."""
        try:
            params = {
                "q": location,
                "appid": self.api_key,
                "units": "metric"
            }
            response = requests.get(self.base_url, params=params)
            response.raise_for_status()
            return response.json()
        except Exception as e:
            logger.error(f"Error fetching weather: {e}")
            return {}

class SlackStatusManager:
    def __init__(self):
        pass

    def update_status(self, status_text: str, emoji: str, expiration: Optional[int] = None) -> bool:
        """Update Slack status."""
        try:
            # We'll implement the actual Slack status update here
            logger.info(f"Updated Slack status: {status_text} {emoji}")
            return True
        except Exception as e:
            logger.error(f"Error updating Slack status: {e}")
            return False

class StatusGenerator:
    def __init__(self, config: Dict):
        self.config = config
        self.templates = config["templates"]
        self.weather_emoji = config["weather_emoji"]
        self.activity_emoji = config["activity_emoji"]

    def generate_status(self, calendar_summary: Dict, weather_data: Dict) -> Dict:
        """Generate appropriate status based on calendar and weather."""
        # We'll implement the status generation logic here
        return {
            "text": "Working from Amsterdam",
            "emoji": "💻",
            "expiration": None
        }

def load_config() -> Dict:
    """Load configuration from YAML file."""
    config_path = os.path.join(os.path.dirname(__file__), "..", "config", "config.yml")
    with open(config_path, "r") as f:
        return yaml.safe_load(f)

def main():
    # Load configuration
    config = load_config()
    logger.info("Configuration loaded successfully")

    # Initialize components
    calendar_handler = CalendarHandler()
    weather_handler = WeatherHandler(os.getenv("OPENWEATHER_API_KEY", ""))
    slack_manager = SlackStatusManager()
    status_generator = StatusGenerator(config)

    try:
        # Get calendar summary
        calendar_summary = calendar_handler.analyze_calendar()
        logger.info("Calendar analysis complete")

        # Get weather data
        weather_data = weather_handler.get_weather(config["user"]["location"])
        logger.info("Weather data retrieved")

        # Generate status
        status = status_generator.generate_status(calendar_summary, weather_data)
        logger.info("Status generated")

        # Update Slack
        success = slack_manager.update_status(
            status["text"],
            status["emoji"],
            status["expiration"]
        )

        if success:
            logger.info("Status updated successfully")
        else:
            logger.error("Failed to update status")

    except Exception as e:
        logger.error(f"Error in main execution: {e}")

if __name__ == "__main__":
    main()
