requirements.txt
pytz>=2023.3
pyyaml>=6.0
requests>=2.31.0
